+++ Fedoraware/Fedoraware-TF2/src/Hooks/Detours/C_BaseEntity_SetAbsVelocity.cpp
@@ -8,9 +8,9 @@ namespace S
 
 MAKE_HOOK(C_BaseEntity_SetAbsVelocity, S::CBaseEntity_SetAbsVelocity(), void, __fastcall, void* ecx, void* edx, const Vec3& vecAbsVelocity)
 {
-	static DWORD dwSetAbsVelocityCall = S::CBasePlayer_PostDataUpdate_SetAbsVelocityCall();
+	static uintptr_t dwSetAbsVelocityCall = S::CBasePlayer_PostDataUpdate_SetAbsVelocityCall();
 
-	if (reinterpret_cast<DWORD>(_ReturnAddress()) == dwSetAbsVelocityCall)
+	if (reinterpret_cast<uintptr_t>(_ReturnAddress()) == dwSetAbsVelocityCall)
 	{
 		if (const auto pBasePlayer = static_cast<CBaseEntity*>(ecx))
 		{
