+++ Fedoraware/Fedoraware-TF2/src/Utils/Pattern/Pattern.h
@@ -3,11 +3,11 @@
 
 class CPattern
 {
-	DWORD FindPattern(DWORD dwAddress, DWORD dwLength, LPCSTR szPattern);
+	uintptr_t FindPattern(uintptr_t dwAddress, DWORD dwLength, LPCSTR szPattern);
 	HMODULE GetModuleHandleSafe(LPCSTR szModuleName);
 
 public:
-	DWORD Find(LPCSTR szModuleName, LPCSTR szPattern);
+	uintptr_t Find(LPCSTR szModuleName, LPCSTR szPattern);
 };
 
 inline CPattern g_Pattern;
